backend/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
web3>=6.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.1
asyncio>=3.4.3

# Data processing and API clients
requests>=2.26.0
pandas>=1.3.0
numpy>=1.21.0

# AI/ML components
google-generativeai>=0.1.0

# Logging and monitoring
loguru>=0.6.0
python-json-logger>=2.0.2

# Testing
pytest>=8.0.0
pytest-asyncio>=1.1.0
pytest-cov>=2.12.0
httpx>=0.18.0

# Security and authentication
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0

# Development tools
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910

# Optional - for enhanced functionality
eth-account>=0.5.7
eth-typing>=2.3.0
eth-utils>=1.10.0